backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
python-multipart==0.0.9
pydantic[email]==2.6.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
aiofiles==23.2.1

# AI/ML
mediapipe==0.10.14
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.4.2
joblib==1.4.0

# NLP
nltk==3.8.1

# Data
pandas==2.2.2

# Testing
pytest==8.1.1
httpx==0.27.0
